backend/requirements.txt
Django==4.2.16
djangorestframework==3.16.1
mongoengine
pymongo==4.15.3
django-cors-headers==4.9.0
python-decouple==3.8
pytz==2024.1
